Multimodal_Fusion/Multimodal_Fusion_Layer/requirements.txt
python-decouple==3.6
pandas==1.5.3
numpy<1.28.0,>=1.21.6
tqdm==4.65.0
redis==5.0.1
-e file:./#egg=multimodal_fusion_layer